Hello, world

+++ src/terminal.h
@@ -78,7 +78,7 @@ class Terminal {
 
   void PrepareBufferClearLine(std::string& buffer) const { buffer += "\x1b[K"; }
 
-  int GetCursorPosition(int& rows, int& cols) const {
+  int GetCursorPosition(size_t& rows, size_t& cols) const {
     char buf[32];
     unsigned int i = 0;
     if (write(STDOUT_FILENO, "\x1b[6n", 4) != 4) return -1;
@@ -89,11 +89,11 @@ class Terminal {
     }
     buf[i] = '\0';
     if (buf[0] != '\x1b' || buf[1] != '[') return -1;
-    if (sscanf(&buf[2], "%d;%d", &rows, &cols) != 2) return -1;
+    if (sscanf(&buf[2], "%zu;%zu", &rows, &cols) != 2) return -1;
     return 0;
   }
 
-  int GetWindowSize(int& rows, int& cols) const {
+  int GetWindowSize(size_t& rows, size_t& cols) const {
     struct winsize ws;
     if (ioctl(STDOUT_FILENO, TIOCGWINSZ, &ws) == -1 || ws.ws_col == 0) {
       if (write(STDOUT_FILENO, "\x1b[999C\x1b[999B", 12) != 12) return -1;
